backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
pydantic>=2.7
pymysql>=1.1
python-dotenv>=1.0
